Fix bugs
git-svn-id: svn://solarus-engine.org/solarus@408 08e91b48-8f19-0410-8fc9-9db6b7bce602

+++ include/FairyMovement.h
@@ -14,7 +14,7 @@ class FairyMovement: public Movement {
   /**
    * Date of the next direction change
    */
-  Uint16 next_direction_change;
+  Uint32 next_direction_change;
 
   void set_random_direction(void);
